Refactor control codes
