+++ camel-canine/canine_controller/ControlUtils/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 3.5)
-project(GaitScheduler LANGUAGES CXX)
+project(ControlUtils LANGUAGES CXX)
 
 include_directories(
         include
@@ -9,6 +9,7 @@ include_directories(
 
 add_library(${PROJECT_NAME} STATIC
         src/Gait.cpp
+        src/SwingLeg.cpp
         )
 
 add_library(canine_controller::${PROJECT_NAME} ALIAS ${PROJECT_NAME})
