+++ camel-canine/canine_controller/PDcontroller/include/PDcontroller/JointPDController.hpp
@@ -10,7 +10,6 @@
 
 #include <camel-tools/CubicTrajectoryGenerator.hpp>
 
-#include "BezierTrajectoryGenerator.hpp"
 
 class JointPDController
 {
@@ -22,13 +21,15 @@ public:
     void InitHomeTrajectory();
     void InitSwingTrajectory();
 
+    void SetControlInput();
+
+private:
     void setTrajectory();
     void computeControlInput();
-    void setControlInput();
+
 
 private:
     CubicTrajectoryGenerator mCubicTrajectoryGen[MOTOR_NUM];
-    BezierTrajectoryGenerator mBezierTrajectoryGen;
 
     double mDesiredP[MOTOR_NUM] = {-0.125, -0.37};
     double mDesiredPosition[MOTOR_NUM];
