+++ camel-canine/canine_controller/convexMPC/CMakeLists.txt
@@ -7,6 +7,7 @@ include_directories(
         ${EIGEN3_INCLUDE_DIRS}
         ${qpOASES_SOURCE_DIR}/include
         ${canine_util_SOURCE_DIR}/include
+        ${ControlUtils_SOURCE_DIR}/include
 )
 link_directories(
         ${qpOASES_SOURCE_DIR}/bin
@@ -14,8 +15,7 @@ link_directories(
 
 add_library(${PROJECT_NAME} STATIC
         src/MPCSolver.cpp
-        src/MPCController.cpp
-        src/SwingLeg.cpp)
+        src/MPCController.cpp)
 
 add_library(canine_controller::${PROJECT_NAME} ALIAS ${PROJECT_NAME})
 target_link_libraries(${PROJECT_NAME} qpOASES)
