+++ camel-canine/canine_fsm/CMakeLists.txt
@@ -10,7 +10,7 @@ include_directories(
         ${canine_gui_SOURCE_DIR}/include
         ${canine_raisim_SOURCE_DIR}/include
         ${canine_util_SOURCE_DIR}/include
-        ${GaitScheduler_SOURCE_DIR}/include
+        ${ControlUtils_SOURCE_DIR}/include
         ${PDcontroller_SOURCE_DIR}/include
         ${convexMPC_SOURCE_DIR}/include
         ${qpOASES_SOURCE_DIR}/include
@@ -35,7 +35,7 @@ SET(CANINE_LIBS
         canine_controller::convexMPC
         canine_controller::ControlMain
         canine_controller::PDcontroller
-        canine_controller::GaitScheduler
+        canine_controller::ControlUtils
         canine_visualizer::canine_gui
         canine_visualizer::canine_raisim
         canine_util
