+++ camel-canine/canine_visualizer/canine_raisim/src/RobotVisualization.cpp
@@ -65,46 +65,46 @@ void RobotVisualization::initRobotPose()
     initialJointPosition.setZero();
     initialJointVelocity.setZero();
 
-//    initialJointPosition[2] = 0.0578;
-//    initialJointPosition[3] = 1.0;
-//
-//    for (int idx=0; idx<4; idx++)
-//    {
-//        initialJointPosition[idx*3+7] = 0.0;
-//        initialJointPosition[idx*3+8] = 2.5;
-//        initialJointPosition[idx*3+9] = -2.9;
-//    }
-
-    // base_x,y,z
-    initialJointPosition[0] = 0.0;
-    initialJointPosition[1] = 0.0;
-    initialJointPosition[2] = 0.37;
-
-    // base_rotation [quaternion]
+    initialJointPosition[2] = 0.0578;
     initialJointPosition[3] = 1.0;
-    initialJointPosition[4] = 0.0;
-    initialJointPosition[5] = 0.0;
-    initialJointPosition[6] = 0.0;
 
-    // FR_hip,thigh,calf
-    initialJointPosition[7] = 0.0;
-    initialJointPosition[8] = 0.7;
-    initialJointPosition[9] = -1.4;
+    for (int idx=0; idx<4; idx++)
+    {
+        initialJointPosition[idx*3+7] = 0.0;
+        initialJointPosition[idx*3+8] = 2.5;
+        initialJointPosition[idx*3+9] = -2.9;
+    }
 
-    // FL_hip,thigh,calf
-    initialJointPosition[10] = -0.0;
-    initialJointPosition[11] = 0.7;
-    initialJointPosition[12] = -1.4;
-
-    // RR_hip,thigh,calf
-    initialJointPosition[13] = 0.0;
-    initialJointPosition[14] = 0.7;
-    initialJointPosition[15] = -1.4;
-
-    // RL_hip,thigh,calf
-    initialJointPosition[16] = -0.0;
-    initialJointPosition[17] = 0.7;
-    initialJointPosition[18] = -1.4;
+//    // base_x,y,z
+//    initialJointPosition[0] = 0.0;
+//    initialJointPosition[1] = 0.0;
+//    initialJointPosition[2] = 0.37;
+//
+//    // base_rotation [quaternion]
+//    initialJointPosition[3] = 1.0;
+//    initialJointPosition[4] = 0.0;
+//    initialJointPosition[5] = 0.0;
+//    initialJointPosition[6] = 0.0;
+//
+//    // FR_hip,thigh,calf
+//    initialJointPosition[7] = 0.0;
+//    initialJointPosition[8] = 0.7;
+//    initialJointPosition[9] = -1.4;
+//
+//    // FL_hip,thigh,calf
+//    initialJointPosition[10] = -0.0;
+//    initialJointPosition[11] = 0.7;
+//    initialJointPosition[12] = -1.4;
+//
+//    // RR_hip,thigh,calf
+//    initialJointPosition[13] = 0.0;
+//    initialJointPosition[14] = 0.7;
+//    initialJointPosition[15] = -1.4;
+//
+//    // RL_hip,thigh,calf
+//    initialJointPosition[16] = -0.0;
+//    initialJointPosition[17] = 0.7;
+//    initialJointPosition[18] = -1.4;
 
     mRobot->setGeneralizedCoordinate(initialJointPosition);
     mRobot->setGeneralizedForce(Eigen::VectorXd::Zero(mRobot->getDOF()));
